Finished basic framework for testing
Completed the basic driving logic for testing, and added some bindings for the operator to invoke these commands.

Also made the PID profiled with a trapezoid profile in the straightTowards command, implemented basic logic for the SelfDriving subsystem(which will be rewritten for the final code) and added some utilities for later use with the priority queue.